[hparams] refactor: split training_args.py into per-algorithm package
Convert the 989-line monolithic training_args.py into a package with
one file per algorithm, improving isolation for parallel development
and reducing merge conflicts as the algorithm count grows.

Structure:
  training_args/
    __init__.py   - re-exports (preserves all existing import paths)
    _base.py      - EvaluationArguments, TrainingArguments, utilities
    _registry.py  - registry dict + get_training_args_class()
    grpo.py       - GRPOTrainingArguments
    nft.py        - NFTTrainingArguments
    awm.py        - AWMTrainingArguments
    dpo.py        - DPOTrainingArguments
    dgpo.py       - DGPOTrainingArguments (extends GRPO)
    crd.py        - CRDTrainingArguments

Zero breaking changes: `from .training_args import X` resolves
identically whether training_args is a .py file or a package __init__.

Also updates ff-new-algorithm skill and sample_lifecycle docs to
reflect the new file paths.
