+++ src/flow_factory/hparams/training_args/_registry.py
@@ -1,0 +1,82 @@
+# Copyright 2026 Jayce-Ping
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Training arguments registry and lookup functions."""
+from __future__ import annotations
+
+import importlib
+from typing import Type, Dict
+
+from ._base import TrainingArguments
+from .grpo import GRPOTrainingArguments
+from .nft import NFTTrainingArguments
+from .awm import AWMTrainingArguments
+from .dpo import DPOTrainingArguments
+from .dgpo import DGPOTrainingArguments
+from .crd import CRDTrainingArguments
+
+
+# ============================================================================
+# Training Arguments Registry
+# ============================================================================
+
+_TRAINING_ARGS_REGISTRY: Dict[str, Type[TrainingArguments]] = {
+    'grpo': GRPOTrainingArguments,
+    'grpo-guard': GRPOTrainingArguments,
+    'nft': NFTTrainingArguments,
+    'awm': AWMTrainingArguments,
+    'dgpo': DGPOTrainingArguments,
+    'dpo': DPOTrainingArguments,
+    'crd': CRDTrainingArguments,
+}
+
+
+def get_training_args_class(identifier: str) -> Type[TrainingArguments]:
+    """
+    Resolve the TrainingArguments subclass for a given trainer type.
+
+    Supports:
+    1. Registry lookup: 'grpo' -> GRPOTrainingArguments
+    2. Direct python path: 'my_package.hparams.CustomTrainingArgs' -> CustomTrainingArgs
+
+    Falls back to base TrainingArguments if lookup fails.
+    """
+    identifier_lower = identifier.lower()
+
+    if identifier_lower in _TRAINING_ARGS_REGISTRY:
+        return _TRAINING_ARGS_REGISTRY[identifier_lower]
+
+    # Try dynamic import (python path like 'my_package.args.CustomArgs')
+    try:
+        module_path, class_name = identifier.rsplit('.', 1)
+        module = importlib.import_module(module_path)
+        cls = getattr(module, class_name)
+        if isinstance(cls, type) and issubclass(cls, TrainingArguments):
+            return cls
+        raise TypeError(
+            f"'{identifier}' resolved to {cls}, which is not a TrainingArguments subclass."
+        )
+    except (ImportError, AttributeError, ValueError, TypeError) as e:
+        raise ImportError(
+            f"Could not resolve TrainingArguments for trainer_type='{identifier}'. "
+            f"Ensure it is either:\n"
+            f"  1. A registered trainer: {list(_TRAINING_ARGS_REGISTRY.keys())}\n"
+            f"  2. A valid python path to a TrainingArguments subclass\n"
+            f"Error: {e}"
+        ) from e
+
+
+def list_registered_training_args() -> Dict[str, Type[TrainingArguments]]:
+    """Get all registered training argument classes."""
+    return _TRAINING_ARGS_REGISTRY.copy()
